+++ src/c/CudaMex/AddImageWith.cpp
@@ -25,7 +25,7 @@ std::string AddImageWith::check( int nlhs, mxArray* plhs[], int nrhs, const mxAr
 	if (!mxIsUint8(prhs[0]) || !mxIsUint8(prhs[1]))
 		return "Images has to be formated as a uint8!";
 
-	int numDims = mxGetNumberOfDimensions(prhs[0]);
+	size_t numDims = mxGetNumberOfDimensions(prhs[0]);
 	if (numDims>3 || numDims<2)
 		return "Image can only be either 2D or 3D!";
